+++ src/instructlab/model/backends/common.py
@@ -19,7 +19,7 @@ LLAMA_CPP = "llama-cpp"
 VLLM = "vllm"
 templates = [
     {
-        "model": "merlinite",
+        "model": "granite",
         "template": "{% for message in messages %}\n{% if message['role'] == 'user' %}\n{{ '<|user|>\n' + message['content'] }}\n{% elif message['role'] == 'system' %}\n{{ '<|system|>\n' + message['content'] }}\n{% elif message['role'] == 'assistant' %}\n{{ '<|assistant|>\n' + message['content'] + eos_token }}\n{% endif %}\n{% if loop.last and add_generation_prompt %}\n{{ '<|assistant|>' }}\n{% endif %}\n{% endfor %}",
     },
     {